+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+## 1.0.0 (2022-09-19)
+
 - Add Phoenix Channels protocol support. ([@palkan][])
 
 - Make `receive` order-independent. ([@palkan][])